move ISSUE_TEMPLATES to templates directory
To prevent release managers from using the GitHub web UI we will be
moving the release templates into a directory that is not used by GH.
