+++ .github/ISSUE_TEMPLATE/release-checklist.md
@@ -1,0 +1,7 @@
+---
+name: Looking for templates? Use ./release checklist open --target-version vX.Y.Z(-rc.W|-pre.N) instead
+about: Create a release checklist for a new version
+title: 'Release checklist for vX.Y.Z'
+labels: ['release']
+assignees: ''
+---
