+++ .github/templates/release_template_rc_branch.md
