+++ testdata/checklist/release_template_minor.md.input
@@ -1,1 +1,1 @@
-../../.github/ISSUE_TEMPLATE/release_template_minor.md
+../../.github/templates/release_template_minor.md
