+++ testdata/checklist/release_template_patch.md.input
@@ -1,1 +1,1 @@
-../../.github/ISSUE_TEMPLATE/release_template_patch.md
+../../.github/templates/release_template_patch.md
