+++ testdata/checklist/release_template_pre_main.md.input
@@ -1,1 +1,1 @@
-../../.github/ISSUE_TEMPLATE/release_template_pre_main.md
+../../.github/templates/release_template_pre_main.md
