+++ testdata/checklist/release_template_rc_branch.md.input
@@ -1,1 +1,1 @@
-../../.github/ISSUE_TEMPLATE/release_template_rc_branch.md
+../../.github/templates/release_template_rc_branch.md
